requirements.txt
Django==1.8.1
coverage==3.7.1
django-angular==0.7.14
django-braces==1.8.0
django-nose==1.4
djangorestframework==3.1.1
model-mommy==1.2.4
nose==1.3.6
six==1.9.0
